fix: pending block & pending block w hash schema fix
add "additionalProperties": false to resolve the conflict
between the pending block schema and the block schema.
this conflict appears in the apis: starknet_getBlockWithTxHashes and starknet_getBlockWithTxs
where we can get either one of the ocjects with a "oneOf" and in fact the block w tx object
is valid for both the block and the pending block schema.